+++ Assets/_Scripts/Player/Weapon.cs
@@ -7,16 +7,25 @@ namespace _Scripts.Player
     public class Weapon : MonoBehaviour
     {
         [SerializeField] private GameObject player;
-        
+
+        private int _damage;
+
         public Camera fpsCam;
         public LayerMask layerMask;
 
-        public int damage = 10;
+        // public int damage = 10;
         public float range = 100f;
+
+        void Start()
+        {
+            // _damage = player.GetComponent<PlayerController>().Stats.Damage;
+        }
 
         // Update is called once per frame
         void Update()
         {
+            _damage = player.GetComponent<PlayerController>().Stats.Damage;
+            
             if (Input.GetButtonDown("Fire1") && player.GetComponent<PlayerController>().GetCurrentPlayerMode() == 0)
             {
                 Debug.Log("Shoot");
@@ -31,14 +40,14 @@ namespace _Scripts.Player
             {
                 if (hit.transform.gameObject.CompareTag("Buildable"))
                 {
-                    hit.transform.gameObject.GetComponent<Buildable>().TakeDamage(damage);
+                    hit.transform.gameObject.GetComponent<Buildable>().TakeDamage(_damage);
                 }
 
                 var enemy = hit.transform.GetComponent<BotController>();
 
                 if (enemy != null)
                 {
-                    enemy.TakeDamage(damage);
+                    enemy.TakeDamage(_damage);
                 }
             }
         }